+++ test/test270.js
@@ -6,12 +6,6 @@ if(typeof exports === 'object') {
 	__dirname = '.';
 };
 
-if(typeof exports == 'object') {
-  var DOMStorage = require("dom-storage")
-  global.localStorage = new DOMStorage("./restest267.json", { strict: false, ws: '' });
-};
-
-
 describe('Test 270 RECORDSET tests', function() {
 
   var emptydata = [];