+++ test/test281.js
@@ -16,7 +16,7 @@ describe('Test 281 UNIQUE Columns (for Meteor-Postgres)', function() {
 
 
   it('2. UNIQUE constraint',function(done){
-    alasql('CREATE TABLE usersTable (id INT, name NVARCHAR(255) NOT NULL UNIQUE)');
+    alasql('CREATE TABLE usersTable (id INT, name NVARCHAR(255) UNIQUE)');
     alasql('INSERT INTO usersTable VALUES (1,"Andrey")');
     alasql('INSERT INTO usersTable VALUES (2,"Kate")');
     done();
@@ -26,6 +26,42 @@ describe('Test 281 UNIQUE Columns (for Meteor-Postgres)', function() {
     assert.throws(function(){
       alasql('INSERT INTO usersTable VALUES (3,"Andrey")');
     },Error);
+    done();
+  });
+
+  it('4. UNIQUE constraint',function(done){
+    alasql('DELETE FROM usersTable WHERE name = "Andrey"');
+    done();
+  });
+
+  it('5. INSERT after deletion',function(done){
+    alasql('INSERT INTO usersTable VALUES (4,"Andrey")');
+    done();
+  });
+
+  it('6. Shoud be the error here with UNIQUE constraint',function(done){
+    assert.throws(function(){
+      alasql('INSERT INTO usersTable VALUES (5,"Andrey")');
+    },Error);
+    done();
+  });
+
+  it('7. Test',function(done){
+    var res = alasql('SELECT * FROM usersTable');
+    assert.deepEqual(res,[ { id: 2, name: 'Kate' }, { id: 4, name: 'Andrey' } ]);
+    done();
+  });
+
+  it('8. Shoud be the error here with UNIQUE constraint',function(done){
+    assert.throws(function(){
+      alasql('UPDATE usersTable SET name = "Andrey" WHERE name = "Kate"');
+    },Error);
+    done();
+  });
+
+  it('9. Test',function(done){
+    var res = alasql('SELECT * FROM usersTable');
+    assert.deepEqual(res,[ { id: 2, name: 'Kate' }, { id: 4, name: 'Andrey' } ]);
     done();
   });